REFACT :: 구현체 분리
